Convert to use /etc/uuid, ephemeral if absent.
This should be 100% predictable and usually all set up before a
client runs. Because we have to test and so on, an ephemeral uuid
will be generated for each process. We may want to add an option
to force ephemeral for testing at some point. It's also possible
we could have boxes that run more than one client service if it's
acting as a centralized poller, but I suspect that will be a
separate piece of code.