[WIP] Add basic plugin class with tests.
This is a work in progress. The test runs but is not successful
because I'm missing something obvious in the plugin class. I'm
going to sleep on it now and look again in the morning.

The plugin executor may actually be useful in other places than
just the client and it codifies a few things so I put it in
Hastur::Plugin::V1.

This commit also adds a bunch of simple test plugins for exercising
the "OK" / exit 0 side. We'll need to add a slew of warning/critical
plugins in the same style.