Check in work on Message and int. test.
The test basically works on Linux on ZMQ 2.1.11. The sleeps
are important though and that's not really great. With 3.1
it doesn't work at all, with messages going to the router,
appearing to get forwarded, then disappearing.

My STDERR debug stuff is still in there for now, will remove
in a later commit when we think we have the behavior locked down.