Switch to to/from instead of route/uuid.
Rather than having a bunch of strings passed around for routing, make things
consistent and have a regular to/from field with some pre-defined recipients
that are the same as routes used to be.

The envelope size is the same, except what was route is now a 128-bit integer
instead of a null-terminated 16-byte string (128 bits).  Because it's cute,
we made the pre-defined routes exactly the old strings packed into a 128-bit
int just like a UUID, so they look similar and are decodable for easy debugging.

Unit tests updated, passing.

Moved the integration test to its own file.