[Hastur::Message] add a few classes/methods
Added some message types that were missing.

Added a few convenience methods the client conversion needs.

Moved some code around to make more sense.

[Hastur::Util] add a basic util module

This should absorb little utilities that are useful all over
and not really specific to a module/class.

Note: I goofed this commit, which is why both files are in it.
This message was updated via amend.