[Hastur::Router] check in new router code
This is the new router code that's totally converted to Hastur::Message
and supports a bunch of the features we've planned. Especially:

* client route caching on-the-fly
* dynamic reconfiguration should work once set up
* not tied to any specific topology

I also went ahead and did the work to keep ZMQ::Message parts that
are for routing entirely in C-space. We're not supposed to mess with
them so copying to Ruby space would likely cause problems. The downside
is we have to remember to .close all of those messages or it'll leak
memory.