[test] another overhaul of integration harness
I made a few oversights in the integration DSL that required a lot
more surgery. The current version is actually running useful tests
(the heartbeat test) so I think we got it.

The tap device that was in HTR::ZeroMQ is gone for now and will reappear
in HT::Tap::ZeroMQ as a dedicated class.

Most of the test libraries had to be touched for this update, since
"action" became "reader" and "writer" so we can pass in lists or blocks
to write to sockets. This should work nicely in the near future with
TCP/UDP/STDIO versions, that will end up a lot less complex than the
ZMQ versions.