[Hastur::Input::JSON] fix json check
The RE was throwing errors on binary data, since it is, you know, binary.

The RE was always a placeholder anyways, and has been replaced by
a simpler check that strips leading/trailing whitespace and simply
checks that the first/last characters are {/}.

Updated tests to new format and added a multi-line test.