[::Input::Collectd] Finish up data conversion.
Finished adding the code to massage collectd data into suitable Hastur
messages for sending on with the same method that statsd and native json
use.

When running against real data on my workstation, I found that some packets
can contain NaN which broke yajl encoding to json and added a really gross
test + workaround.  A test has been added to check for this in the future.

Ran for a half hour against real data on my box afterwards and everything looks
solid.