Add HMAC and Traceroute features to message.
Hastur::Envelope can now transmit and compute HMAC authentication
for the data payload. It's not totally integrated yet, but it is
tested.

Hastur::Envelope can also now append an unbounded list of UUID's
to the tail of the envelope so we can track the flow through Hastur.
Most of the time, there will only be one additional 128-bit UUID
on each envelope, added at the router.

Hastur::Message::Base.send can take a :secret argument which sets
up the HMAC in the envelope just before sending. We'll want to come
up with a clever way to manage the secret so we don't have to pass
it around everywhere. Either a singleton/class variable or some kind
of factory (*shudder*).

Tests are updated and passing. Most only needed constants for size
updated.