HUGE REFACTOR: split ::Message, rework ::Router
I finally tracked down the bug(s) in the router that were dropping acks.  The most important
part of the fix to the router was adding a type field to the envelope that we can use to
automatically reconstitute message objects rather than abusing the to/from fields (which we
already abuse a little for to => <route>).

The change over to type involved touching a lot of pieces of envelope, and exposed some
weaknesses and confusing parts of Hastur::Message. As part of all the changes I decided
to split all the subtypes out to one file per class.

The message test was split into message / envelope, so the message subtype tests are still
in one file for now, and envelope is in its own.

test/integration/ack_test.rb was the big driver I hammered on while working on these changes.
This test now works as expected.  I found a couple bugs in Nodule and fixed them, so 0.0.5 is now
required. There is now a more sane way to synchronize test startup using a message counter. Check
out ack_test.rb.

A bunch of small things had to be touched all over to convert to the new, clearer id/symbol/uuid
representation of classes/routes. There are new class methods in Hastur::Message for converting
from type_id, symbol (e.g. :stat), and route_uuid to a class name. The static hashes are in place
to support other methods as they become necessary.