[ALL THE THINGS] remove magic routing addresses
Since we added the 8-bit type into the message envelope, most of the
things we were using the magical route UUID's for became redundant
and confusing.

This commit removes the magic UUID's and switches routing over to use
types or strict to/from UUID's without magical addresses. In many places
we now use an all-zero :to or :from address for "I don't know, so figure
it out for me". It would be packed as all zeroes anyways, so I feel like
requiring the address and setting it to 0 is nice & accurate compared to
defaulting in some places (e.g. Envelope still always requires an address
and never tries to make one up for you).