[test] get query server test working partially
The second REST request still fails and is commented out.

At this point, the startup works OK, but I'm not happy with having to
sleep 6 seconds to make sure data is flushed to Cassandra. I'm thinking
about removing the real clients and generating heartbeats directly with
ack enabled so we can know for sure when they have hit the database,
since the acks only send after commit.