[test] try to get the query server test working
Something is weird with cassandra-cli during the bootstrap. If I
move UUIDDay to the top of the schema file it loads enough for the
test to run OK. Otherwise, cli craps out about half way through the
CF definitions leaving "org.apache.thrift.transport.TTransportException"
in STDERR.

So this version is still broken, but it's less broken.