[test] query server test works on my machine now
Turned out it was a really silly issue all along (besides the cassandra
start/stop which was fixed elsewhere).

The heartbeat resource was being connected to by both Nodule::ZeroMQ
and the cassandra sink. So ZeroMQ just did the right thing and round
robin'ed across them, making it look like half of the messages were
disappearing. I spent a stupid amount of time trying to figure this out ...