Add new package build tasks.
I started adding a bunch of junk to the shell script and decided to
rewrite it as rake tasks. It turns out to be much larger than the
shell script, but should be a lot easier to work with over time since
most things are in individual tasks. This makes incremental build debugging
much easier since you can just run specific tasks rather than editing the file.