[util] address CR, fix bugs
There were bugs in connect() and bind() I found while updating the docs.

Adjusted a bunch of docs to be more correct.

Moved zmq_error to be private at the bottom of the file: Hastur::Util
is the only user of it and it doesn't save any line space anywhere else.