[rest] remove /data/ from paths, cleanups, better errors, etc..
I know this one might be a little contentious. I couldn't remember why we wanted
/data/ in some of the paths but not others and it was annoying to work with so I
removed it. I also took some redundant paths with it so things are looking good to
my eyes.

We can add /data/ back if/when Noah reminds me why we did it in the first place.

* simplified & reduced most routes, they're consistent now
* deserialize messages by default so JSON to the client only needs to be deserialized once
* change the "stat" name to "metric"
* add info_ohai to the info list in TYPES
* add a couple constants for flipping values/messages and five_minutes/one_day
* add some configuration so sinatra doesn't barf html
* shortened foo_bar names to be simpler, may still need tweaking
* removed cut & paste yard and put an @see instead
* /node path returns more resources with sane defaults for values/messages