[rest] reorder hashes, add sample=N parameter
"sample=<int>" sets a maximum number of samples to return, it does _not_ roll up, it figures out how
many entries to skip and selects only every (total / samples) point out of the data. This works with
either of the above options or all by itself.

While working on the sample= stuff and rickshaw graphs, we were seeing unusual spikes in the data. This
turned out to be out-of-order entries in the data. Without tracking it down, I've added a block to sort
the timestamps and reinsert them into a new hash on every run.  We should probably try to get the Cassandra
queries to always return correct ranges and merge them correctly, but that's probably going to be pretty
hard to track down and fix, so we'll defer.