[rest] add first pass at transformation proxy
This will absorb all the extra formats and other miscellaneous work
that isn't purely retrieval.

It's using Goliath at the moment, because I was already experimenting
with it and it'll make a nice place to hook up websockets down the road.

Most of this code except for the request handler at the top is copy/paste
from retrieval.rb, which I'm going to strip of all this code momentarily.