[logging] enhance logging, make inserts consistent
Something was failing on insert but no longer is. One bug was in the rescue
that tried to call .to_hash on a nil value, causing a thread crash in hastur-core.rb.

The :LookupByKey inserts were not using the insert options and relying on Cassandra
defaults. This could in theory be more efficient (since lookup tables could be RF=1)
but consistency is to be preferred.