[envelope] show String.inspect on UUID errors
We're occasionally getting valid_uuid? failures on the production core
and what it shows in the log is actually a valid UUID if I copy/paste
and test.

This switches the log message to call inspect which should escape any
non-ascii hiding in there, which is what I currently suspect.