[agent] add /etc/cnames to regs, refactor ohai/reg a little
/etc/cnames is authoritative for the human-facing name at Ooyala. Read it and report it as
it may not always sync up with DNS, but is the one true way to get the most-preferred name.

Refactor the uname stuff to run inside begin/rescue in case it errors.

Adjust timers for reg / ohai to be more accurate than the silly number before. Make Ohai
report after 60 seconds of running as it's pretty heavy and seems to fail sporadically during
startup turmoil.