[retrieval] work around weird duplicate timestamp error in /message
Jigish saw some oddness when querying multiple names for messages where the same
message would be deserialized twice, so the second one would be a stringified ruby
hash instead of json, throwing an exception.

This works around it for now by stupidly checking the type of the data before trying
to deserialize and assume that a Hash is an already deserialized entity.