[time_util] add month alignment, improve docs/errors
Months are not really viable intervals, so they're in their own method
with its own docs & tests.

I originally had usec_aligned_chunks call through to usec_aligned_months for :month
but I felt it gives the wrong impression since a month chunk is a totally different
animal than hour/day/week which we can mostly assign fixed durations to.

As the docs say, month chunks are defined as aligned to the 1st of the month at midnight UTC.

Weeks may seem similar, but we assign them to be 7 * 24 hours rather than basing it on the calender
whereas months require calendar math.

This is intended to be used for rollup bucket timestamps.