[retrieval] remove consistency parameter, force LOCAL_QUORUM
This parameter wasn't terribly useful since you'd have to know the constants
to use it, which is janky. I was sometimes seeing inconsistent results when
repeating requests, which was probably due to reading with consistency=>1,
which is a bad idea unless we replicate everything, which we don't, so
don't do that anymore.