[cassandra/schema] light refactoring
Merge two nearly identical blocks into the row_keys method call that generates
a list of row keys from a list of uuids, a bucket size, and timestamp. This method
is documented.

Use a local constant for consistency that's easier to keep in sync with tests that
will break if it changes.

Add a note in lookup_by_key to call out the fact that its time bucket is not currently
configurable.