[retrieval] enable lookup agg functions, allow narrowing /api/name
The cname(), fqdn(), and hostname() functions require access to Cassandra
and while they're at it, it's a really good idea to pass in the user-specified
timestamps in case a GUID doesn't exist outside that time range.

Added a @logger.error in hastur_error so we can see exceptions in syslog.

Added a quick change to allow specifying ?uuid= or ?type= on /api/name/:name
queries to allow narrowing down to a specific type on name searches.