add the first (obviously) heuristic function, unrollup()
http://en.wikipedia.org/wiki/Heuristic#Computer_science

This function will likely do awful things on the wrong kind of data, but is necessary
for most operation systems counters. Tools like RRD try really hard to hide rollovers
whereas we keep them intact and act on them at the end of the chain via functions like
this one.