Add a Grape-based REST API V1.
This is pretty much a refactor of the retrieval service, but in all-new files. The retrieval
service is unchanged so far.

I got tired of hacking around Sinatra weirdness and have been wanting to try
Grape, so I did a quick conversion that turned into a nice refactoring of
the retrieval service.

The YARD still needs some massaging, but it's all there and it's time for me to sleep.

Major changes:
  * uses /v1 instead of /api (so we can run both!)
  * helpers are in their own file
  * constants are in their own file
  * "format" has been renamed to "kind" to not collide with Grape's :format parameter
  * a bunch of the /node/:uuid/* routes are merged into one multi-route

TODO:
  * better error handler block that gives more information over REST than an emtpy 404
  * figure out why YARD isn't rendering the nested resource routes, clean up YARD errors
  * the top-level /v1 route is broken (see @todo in v1.rb)
  * pretty print is not available with Grape yet (I have patches to send them)
  * need to set up Rack::JSONP to replace former JSONP hack
  * use Grape's parameter checking everywhere
  * convert retrieval service's unit test