Run the agent as a user in production.
Plugins kinda needed root, but what's left after they're gone doesn't.

The after-install script for the Debian package installs the role-hastur user
automatically and bluepill will now run as that user.

We pretty much always test the agent as a non-root user, so this is a fairly
safe change. The package is the main thing to test, which I will do manually
during the next package build cycle.