Convert Cassandra schema to snake_case from CamelCase.
This makes the data acecssible (even if unreadable) via CQL. I wanted this done
before we fire up the new cluster since we're migrating all the data anyways.

Tests pass on my box and I fired up the whole infrastructure and it works.

Also cleaned up the whole "stat_*" thing so now it's "*_value" which matches all
the other CF's.