[agent] make the hastur api usage a little clearer
IMO this is a lot clearer - I had to stop and think a bit about what was
happening with "require hastur". Most of the libraries used by the agent
have their own require hastur/api line but this still feels better. Should
be a noop change at the end of the day.