[agent] fix signal handler to do clean shutdown
Previously, the signal handler was calling agent.shutdown, which sets the
run loop to false and then shuts down all the sockets. This usually happened
inside of a select() call which results in exceptions. One solution is to update
the select loop to do the right thing, which we should also do, but for now the
other correct solution is to let the run loop exit naturally, then do the shutdown.