[monitoring] add hastur-sentinal.rb
Yup, it's nasty. Yup, it works and pages me when crap breaks.

From the file:

hastur-sentinel - a fairly dirty full-loop monitoring daemon for hastur production

This should catch most errors where we're losing messages due to thrift_client
breakage, broken hastur-core, or any connectivity issues between all the systems.

Starts up an agent for each router specified on the command line and sends messages
through that agent, then hits the retrieval service to verify that the message was
written to the database and is readable.

This is also handy with really low wait/stall values to test various message loss
scenarios. I already used it to show that the agent HWM is working as expected and
resending messages when the router comes back up. This is why it prints the curl
command on failures.

TODO: rewrite it. Globals, nasty hackery on alerts, needs retries, etc.
Good Parts:
 * the approach seems to work
 * pagerduty integration
 * can run in really tight loops and saves metrics to Hastur